requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
attrs==25.3.0
blinker==1.9.0
click==8.2.1
discord.py==2.5.2
enum-tools==0.13.0
Flask==3.1.1
frozenlist==1.7.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.4
propcache==0.3.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
six==1.17.0
twitchAPI==4.5.0
typing_extensions==4.14.1
Werkzeug==3.1.3
yarl==1.20.1
